+++ src/earn/poolInfoScreen/EarnPoolInfoScreen.tsx
@@ -16,8 +16,8 @@ import { LabelWithInfo } from 'src/components/LabelWithInfo'
 import { formatValueToDisplay } from 'src/components/TokenDisplay'
 import TokenIcon, { IconSize } from 'src/components/TokenIcon'
 import Touchable from 'src/components/Touchable'
-import BeforeDepositBottomSheet from 'src/earn/BeforeDepositBottomSheet'
 import { useDepositEntrypointInfo } from 'src/earn/hooks'
+import BeforeDepositBottomSheet from 'src/earn/poolInfoScreen/BeforeDepositBottomSheet'
 import { SafetyCard } from 'src/earn/SafetyCard'
 import { getEarnPositionBalanceValues } from 'src/earn/utils'
 import WithdrawBottomSheet from 'src/earn/WithdrawBottomSheet'
